[UPD] add a wizard to create a project directly from the BRs
