+++ business_requirement_project/__openerp__.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+# © 2016 Elico Corp (https://www.elico-corp.com).
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 {
     'name': 'Business Requirement - Project',
     'category': 'Business Requirements Management',
