+++ business_requirement_project/models/__init__.py
@@ -1,2 +1,5 @@
 # -*- coding: utf-8 -*-
+# © 2016 Elico Corp (https://www.elico-corp.com).
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 from . import business
+from . import project
