+++ business_requirement_project/wizard/generate_projects.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
+# © 2016 Elico Corp (https://www.elico-corp.com).
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 from openerp import models, fields, api
 from openerp.tools.translate import _
 
@@ -20,6 +22,15 @@ class BrGenerateProjects(models.TransientModel):
     )
     for_br = fields.Boolean('Create sub-projects for Business requirements')
     for_deliverable = fields.Boolean('Create sub-projects for Deliverables')
+    for_childs = fields.Boolean(
+        'Create sub-projects for Child Business requirements')
+    br_ids = fields.Many2many(
+        string='Business requirements',
+        comodel_name='business.requirement',
+        relation='wizard_br_rel',
+        column1='wizard_id',
+        column2='br_id'
+    )
 
     @api.multi
     def wizard_view(self):
@@ -46,18 +57,16 @@ class BrGenerateProjects(models.TransientModel):
         project_ids = []
         parent_project = self.project_id
         if not self.for_br and not self.for_deliverable:
-            for br in parent_project.br_ids:
+            for br in self.br_ids:
                 lines = [
                     line.resource_ids for line in br.deliverable_lines
                     if line.resource_ids
                 ]
                 self.create_project_task(lines, parent_project.id, task_ids)
         else:
-            for br in parent_project.br_ids:
+            for br in self.br_ids:
                 self.generate_br_projects(
                     parent_project, br, project_ids, task_ids)
-        import pdb
-        pdb.set_trace()
         if project_ids:
             ids = ['%s' % x for x in project_ids]
             res_model = 'project.project'
@@ -110,11 +119,10 @@ class BrGenerateProjects(models.TransientModel):
             self.generate_deliverable_projects(
                 line_parent, br.deliverable_lines, project_ids, task_ids)
 
-        for child_br in br.business_requirement_ids:
-            if child_br.project_id and \
-                    (child_br.project_id.id == parent_project.id):
-                continue
-            self.generate_br_projects(parent_project, child_br, project_ids)
+        if self.for_childs:
+            for child_br in br.business_requirement_ids:
+                self.generate_br_projects(
+                    parent_project, child_br, project_ids, task_ids)
 
     @api.multi
     def generate_deliverable_projects(
@@ -134,7 +142,7 @@ class BrGenerateProjects(models.TransientModel):
 
     @api.multi
     def _prepare_project_vals(self, br, parent):
-        project = {
+        vals = {
             'name': br.description,
             'parent_id': parent.id,
             'partner_id': parent.partner_id.id,
@@ -144,7 +152,7 @@ class BrGenerateProjects(models.TransientModel):
             'user_id': parent.user_id.id,
             'origin': '%s.%s' % (br._name, br.id)
         }
-        return project
+        return vals
 
     @api.multi
     def _prepare_project_task(self, line, project_id):
@@ -153,7 +161,7 @@ class BrGenerateProjects(models.TransientModel):
         product_uom_obj = self.env['product.uom']
         qty = product_uom_obj._compute_qty(
             line.uom_id.id, line.qty, default_uom)
-        task = {
+        vals = {
             'name': line.task_name,
             'description': line.description,
             'sequence': line.sequence,
@@ -161,7 +169,7 @@ class BrGenerateProjects(models.TransientModel):
             'planned_hours': qty,
             'br_resource_id': line.id,
         }
-        return task
+        return vals
 
     @api.multi
     def create_project_task(self, resource_lines, project_id, task_ids=[]):
